test: React 17 & 18 testing
