+++ .github/workflows/test.yml
@@ -30,3 +30,5 @@ jobs:
         run: yarn build
       - name: Test
         run: yarn test
+      - name: Test React 17
+        run: yarn test-17
