+++ package.json
@@ -10,6 +10,7 @@
     "start-test-harness": "yarn workspace browser-test start-harness",
     "test-dev": "yarn workspace browser-test test-dev",
     "test": "yarn workspace browser-test test",
+    "test-17": "yarn workspace browser-test test-17",
     "start-docs": "yarn workspace docs develop",
     "build-docs": "yarn workspace docs build",
     "update-yarn-ts": "yarn dlx @yarnpkg/sdks vscode"
