+++ packages/browser-test/test-harness-17/.gitignore
@@ -1,0 +1,1 @@
+../test-harness/.gitignore
