+++ packages/browser-test/test-harness-17/craco.config.js
@@ -1,0 +1,9 @@
+function overrideWebpackConfig({ webpackConfig }) {
+  webpackConfig.resolve.symlinks = false;
+
+  return webpackConfig;
+}
+
+module.exports = {
+  plugins: [{ plugin: { overrideWebpackConfig }, options: {} }],
+};
