+++ packages/browser-test/test-harness-17/public/index.html
@@ -1,0 +1,1 @@
+../../test-harness/public/index.html
