+++ packages/browser-test/test-harness-17/src/index.tsx
@@ -1,0 +1,4 @@
+import ReactDOM from 'react-dom';
+import { start } from './shared/main';
+
+start(ReactDOM.render);
