+++ packages/browser-test/test-harness-17/src/shared
@@ -1,0 +1,1 @@
+../../test-harness/src/shared
