+++ packages/browser-test/test-harness-17/tsconfig.json
@@ -1,0 +1,1 @@
+../test-harness/tsconfig.json
