+++ packages/docs/package.json
@@ -8,7 +8,7 @@
     "@mdx-js/react": "^1.6.22",
     "@types/github-slugger": "^1.3.0",
     "@types/mdx-js__react": "^1.5.5",
-    "@types/react": "^17.0.44",
+    "@types/react": "^18.0.1",
     "@types/react-helmet": "^6.1.5",
     "@types/styled-components": "^5.1.25",
     "babel-plugin-styled-components": "^2.0.7",
@@ -22,8 +22,8 @@
     "github-slugger": "^1.4.0",
     "polished": "^4.2.2",
     "prism-react-renderer": "^1.3.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-helmet": "^6.1.0",
     "react-responsive-pagination": "workspace:packages/react-responsive-pagination",
     "rehype-slug": "^4.0.1",
